Update patch set 1
Patch Set 1: Code-Review+1

Tested and working 100% as designed. We couldn't get it to display consistently using various eri.xml workarounds natively. The only way it used to work on a regular basis was when you could still put a custom eri.xml in /data, which no longer works, and is not wanted anyways. This simply changes the "CDMA Roaming Indicator Strings (non ERI)" as stated above. If eri gets sorted out and we can get this to work 100% of the time pulling from eri.xml then this could be pulled if needed down the road. Verizon's text for Roaming Indicator "3" is actually "Network Extender" and not "Out of Neighborhood".

Patch-set: 1