Update patch set 5
Patch Set 5:

I was looking at http://review.cyanogenmod.org/#/c/60661/ and notice there isn't a list for "preferred_network_mode_values". Just thinking will this be the cause of the problem as the menu list aren't define properly to the function it should be selecting, causing the fc.

Patch-set: 5