Update patch set 2
Patch Set 2:

(4 comments)

I don't completely disagree with this but at the same time I'm unclear of the benefits of this exactly (due to my lack of research). I'd want to test it before merging and also probably look at commonizing some of it with jf and possibly hlte.

Steve may have an opinion as well so I'm going to add him to this commit too.

Patch-set: 2