Update patch set 2
Patch Set 2:

Very poor performance in my testing. Works fine when the phone is plugged in. Does not wake from sleep usually when not plugged in (usually ends up hard resetting when I hold power down too long) and sometimes when it's supposed to sleep it enables airplane mode (icon changes in status bar) and the UI becomes unresponsive, but it doesn't actually sleep.

Patch-set: 2